refactor(pytest|consume): re-structure and refine pytest & consume commands. (#714)
