+++ setup.cfg
@@ -78,10 +78,10 @@ pytest_plugins =
 
 [options.entry_points]
 console_scripts =
-    fill = cli.pytest_commands:fill
-    tf = cli.pytest_commands:tf
+    fill = cli.pytest_commands.fill:fill
+    tf = cli.pytest_commands.fill:tf
     checkfixtures = cli.check_fixtures:check_fixtures
-    consume = cli.pytest_commands:consume
+    consume = cli.pytest_commands.consume:consume
     genindex = cli.gen_index:generate_fixtures_index_cli
     gentest = cli.gentest:make_test
     pyspelling_soft_fail = cli.tox_helpers:pyspelling
