+++ src/pytest_plugins/help/help.py
@@ -52,10 +52,7 @@ def show_test_help(config):
             "defining debug",
         ]
     elif pytest_ini.name in [
-        "pytest-consume-all.ini",
-        "pytest-consume-direct.ini",
-        "pytest-consume-rlp.ini",
-        "pytest-consume-engine.ini",
+        "pytest-consume.ini",
     ]:
         test_group_substrings = [
             "execution-spec-tests",
